requirements-dev.txt
lxml==4.4.2
twstock==1.3.1
python-telegram-bot==12.3.0
pytz==2019.3
numpy==1.18.1
pandas==0.25.3